+++ orix/sampling/__init__.py
@@ -16,18 +16,31 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Generation of grids on *S2* or *SO(3)* (rotation space)."""
+"""Generation of grids on *S2* (vectors) or *SO(3)* (rotations)."""
 
 from orix.sampling.sample_generators import get_sample_fundamental, get_sample_local
 from orix.sampling.SO3_sampling import uniform_SO3_sample
-from orix.sampling.S2_sampling import sample_S2_cube_mesh, sample_S2_uv_mesh
+from orix.sampling.S2_sampling import (
+    sample_S2,
+    sample_S2_cube_mesh,
+    sample_S2_hexagonal_mesh,
+    sample_S2_icosahedral_mesh,
+    sample_S2_random_mesh,
+    sample_S2_uv_mesh,
+    sampling_methods as sample_S2_methods,
+)
 
 
 # Lists what will be imported when calling "from orix.sampling import *"
 __all__ = [
     "get_sample_fundamental",
     "get_sample_local",
-    "sample_S2_cube_mesh",
-    "sample_S2_uv_mesh",
+    "sample_S2",
+    "sample_S2_methods",
     "uniform_SO3_sample",
+    "sample_S2_cube_mesh",
+    "sample_S2_hexagonal_mesh",
+    "sample_S2_icosahedral_mesh",
+    "sample_S2_random_mesh",
+    "sample_S2_uv_mesh",
 ]