+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Added
 -----
 - Support for type hints has been introduced and a section on this topic has been added
   to the contributing guide.
+- Seven methods for sampling unit vectors from regular grids on *S2* via
+  ``orix.sampling.sample_S2()``.
 - Plotting IPF color maps on axes now possible on through
   `InversePoleFigurePlot.plot_ipf_color_key()`.